backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
PyJWT==2.8.0
Cryptography==3.4.7
Boto3==1.26.0
python-dotenv==0.19.0
SQLAlchemy==1.4.49
pytest==8.0.2
Google-GenerativeAI==0.3.2
botocore==1.29.0
six==1.16.0
